Fixed bugs in save_brain_data.py

+++ brain_data/save_brain_data.py
@@ -11,14 +11,16 @@ import time
 import numpy as np
 import matplotlib.pyplot as plt
 import pickle
-from in_progress import cont_fourier_trans as analysis
+from analysis_tools import get_power_spectrum, get_rms_voltage
 from Adafruit import ADS1x15
 
-ACQTIME = 5
+ACQTIME = 1
 SPS = 920 #Samples per second to collect data. Options: 128, 250, 490, 920, 1600, 2400, 3300.
 VRANGE = 6144 #Full range scale in mV. Options: 256, 512, 1024, 2048, 4096, 6144.
 nsamples = int(ACQTIME*SPS)
 sinterval = 1.0/SPS
+freq_min = 8 #min freq of alpha waves
+freq_max = 12 #max freq of alpha waves
 
 print()
 print('Initializing ADC...')
@@ -26,33 +28,45 @@ print()
 
 adc = ADS1x15()
 
-save_path = input('Please input folder name to make or save to:')
+"""
+Ask user to specify what folder to save in
+"""
+while True:
+    person_name = input('Please enter your name (hak/ryan/ruining):')
+    if person_name in ['hak', 'ryan', 'ruining']:
+        break
+    else:
+        print('Please enter a correct formatted name')
+        
+save_folder = input('Please input folder name to make or save to:')
+save_path = os.path.join(person_name, save_folder)
 if not os.path.isdir(save_path): #Check if exp_dir is already a directory
     try:
         os.mkdir(save_path)
-        os.mkdir(os.path.join(save_path,'/relaxed'))
-        os.mkdir(os.path.join(save_path,'/concentrated'))
         print("Successfully created folders")
     except OSError:
         print("Creation of the folder %s failed" % save_path)
 else:
     print("Successfully connected to the folder %s " % save_path)
-
+    
 while True:
-    wavetype = input('Record relaxed or concentrated waves? Type r or c.)')
+    wavetype = input('Record relaxed or concentrated waves? Type (r/c)')
     if wavetype in ['r','c']:
         break   
     else:
         print('Please type r or c')                   
 if wavetype == 'r':
-    exp_file = os.path.join(save_path,'/relaxed.py')
+    exp_file = os.path.join(save_path,'relaxed.pickle')
 elif wavetype == 'c':
-    exp_file = os.mkdir(os.path.join(save_path,'/concentrated.py'))
-                   
+    exp_file = os.path.join(save_path,'/concentrated.pickle')
+
+"""
+Continue to collect brain wave data until user specifies stop
+"""
 end_program = 'n'
 while end_program != 'y': #Loops every time user records data
 
-    input('Press <Enter> to start %.1f s data acquisition...' % recordtime)
+    input('Press <Enter> to start %.1f s data acquisition...' % ACQTIME)
     print()
     adc.startContinuousDifferentialConversion(2, 3, pga=VRANGE, sps=SPS)
     time_series = np.zeros(nsamples, 'float')
@@ -64,8 +78,9 @@ while end_program != 'y': #Loops every time user records data
         time_series[i] -= 3.3 #ADC ground is 3.3 volts above circuit ground
         while (time.perf_counter() - st) <= sinterval:
             pass
-    ps = analysis.get_power_spectrum(time_series)
-    rms = analysis.get_rms_voltage(ps, freq_min, freq_max)
+    freq = np.fft.fftfreq(nsamples, d=1.0/SPS)
+    ps = get_power_spectrum(time_series)
+    rms = get_rms_voltage(ps, freq_min, freq_max, freq, nsamples)
     print('RMS of Alpha Wave Voltage: ', rms)
     
     f1, ax1 = plt.subplots()
@@ -74,31 +89,41 @@ while end_program != 'y': #Loops every time user records data
     ax1.set(xlabel='Time (s)', ylabel='Voltage', title='Raw Signal')    
     f1.show()
     
-    freqs = np.fft.fftfreq(len(ps), d=1.0/SPS)
     f2, ax2 =plt.subplots()
     plt.xlim(0,200)
-    ax2.plot(freqs, ps)
+    ax2.plot(freq, ps)
     ax2.set(xlabel='Frequency (Hz)', ylabel='Power Spectrum', title='Power Spectrum of FFT')    
     f2.show()
     
     while True:
         save = input('Save this last run? (y/n)')
         if save == 'y':
-            file = open(exp_file, 'wb')
-            pickle.dump(time_series, file)
-            file.close()
+            if os.path.exists(exp_file): #adds to existing file
+                file = open(exp_file, 'rb')
+                brain_data = pickle.load(file)
+                file.close()
+                brain_data.append(time_series)
+                file = open(exp_file, 'wb')
+                pickle.dump(brain_data, file)
+                file.close()
+            else: #creates new file to save
+                brain_data = [time_series]
+                file = open(exp_file, 'wb')
+                pickle.dump(brain_data, file)
+                file.close()
             break            
         elif save == 'n':
             break
         else:
             print('Please type y or n.')
+    plt.close('all')
     while True:
         end_program = input('Do you want to end program? (y/n)')
         if end_program in ['y', 'n']:
             break
         else:
             print('Please enter \'y\' or \'n\' ')
-    print('\n \n \n')
+    print('\n')
